Improving Spread
git-svn-id: svn://activequant.com/svn/repositories/aq2o/trunk@296 01b8edfa-0b2f-4ad0-b428-228ee2dd3b63

+++ src/main/java/com/activequant/dao/mybatis/GenericMapperDao.java
@@ -171,7 +171,7 @@ public class GenericMapperDao<T extends PersistentEntity> {
 		} else if (value instanceof String) {
 			gr = new GenericRow(createdTimeStamp, id, key, null, null, (String) value);
 		} else if (value instanceof Integer) {
-			gr = new GenericRow(createdTimeStamp, id, key, (Long) value, null, null);
+			gr = new GenericRow(createdTimeStamp, id, key, ((Integer) value).longValue(), null, null);
 		} else if (value instanceof Long) {
 			gr = new GenericRow(createdTimeStamp, id, key, (Long) value, null, null);
 		}
